Rename vit_pytorch to vit
